requirements.txt
pytest
playwright
allure-pytest
python-dotenv
pytest-playwright